requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
Django==4.2.5
django-cors-headers==4.2.0
django-mssql-backend-azure==2.9.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
msal==1.24.0
mssql-django==1.3
pycparser==2.21
PyJWT==2.8.0
pyodbc==4.0.39
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
